Added a separate ConnectorDetailPageQueryService
