+++ extensions/broker-server/src/main/java/de/sovity/edc/ext/brokerserver/dao/pages/connector/ConnectorListPageQueryService.java
@@ -1,0 +1,4 @@
+package de.sovity.edc.ext.brokerserver.dao.pages.connector;
+
+public class ConnectorListPageQueryService {
+}
